+++ geo/metadata.go
@@ -1,4 +1,4 @@
-package parser
+package geo
 
 import (
 	"bytes"
@@ -9,15 +9,10 @@ import (
 	"net/http"
 	"net/url"
 	"os"
-	"reflect"
 	"strconv"
-	"strings"
 	"time"
 
-	"cloud.google.com/go/bigquery"
-
 	"github.com/m-lab/etl/metrics"
-	"github.com/m-lab/etl/schema"
 	"github.com/prometheus/client_golang/prometheus"
 )
 
@@ -38,15 +33,15 @@ var BatchURL = AnnotatorURL + "/batch_annotate"
 // same length. It will then make a call to the batch annotator, using
 // the ip addresses and the timestamp. Then, it uses that data to fill
 // in the structs pointed to by the slice of GeolocationIP pointers.
-func FetchGeoAnnotations(ips []string, timestamp time.Time, geoDest []*schema.GeolocationIP) {
-	reqData := make([]schema.RequestData, 0, len(ips))
+func FetchGeoAnnotations(ips []string, timestamp time.Time, geoDest []*GeolocationIP) {
+	reqData := make([]RequestData, 0, len(ips))
 	for _, ip := range ips {
 		if ip == "" {
 			metrics.AnnotationErrorCount.With(prometheus.
 				Labels{"source": "Empty IP Address!!!"}).Inc()
 			continue
 		}
-		reqData = append(reqData, schema.RequestData{ip, 0, timestamp})
+		reqData = append(reqData, RequestData{ip, 0, timestamp})
 	}
 	annotationData := GetBatchMetaData(BatchURL, reqData)
 	timeString := strconv.FormatInt(timestamp.Unix(), 36)
@@ -62,11 +57,12 @@ func FetchGeoAnnotations(ips []string, timestamp time.Time, geoDest []*schema.Ge
 	}
 }
 
+/* These should be in ss.go and pt.go
 // AddMetaDataSSConnSpec takes a pointer to a
 // Web100ConnectionSpecification struct and a timestamp. With these,
 // it will fetch the appropriate metadata and add it to the hop struct
 // referenced by the pointer.
-func AddMetaDataSSConnSpec(spec *schema.Web100ConnectionSpecification, timestamp time.Time) {
+func AddMetaDataSSConnSpec([]string, server *GeolocationIP, timestamp time.Time) {
 	if spec == nil {
 		metrics.AnnotationErrorCount.With(prometheus.
 			Labels{"source": "SS ConnSpec was nil!!!"}).Inc()
@@ -81,7 +77,7 @@ func AddMetaDataSSConnSpec(spec *schema.Web100ConnectionSpecification, timestamp
 	}(timerStart)
 
 	ipSlice := []string{spec.Local_ip, spec.Remote_ip}
-	geoSlice := []*schema.GeolocationIP{&spec.Local_geolocation, &spec.Remote_geolocation}
+	geoSlice := []*GeolocationIP{server, client}
 	FetchGeoAnnotations(ipSlice, timestamp, geoSlice)
 }
 
@@ -103,7 +99,7 @@ func AddMetaDataPTConnSpec(spec *schema.MLabConnectionSpecification, timestamp t
 			Observe(float64(time.Since(tStart).Nanoseconds()))
 	}(timerStart)
 	ipSlice := []string{spec.Server_ip, spec.Client_ip}
-	geoSlice := []*schema.GeolocationIP{&spec.Server_geolocation, &spec.Client_geolocation}
+	geoSlice := []*GeolocationIP{&spec.Server_geolocation, &spec.Client_geolocation}
 	FetchGeoAnnotations(ipSlice, timestamp, geoSlice)
 }
 
@@ -126,7 +122,7 @@ func AddMetaDataPTHopBatch(hops []*schema.ParisTracerouteHop, timestamp time.Tim
 // AnnotatePTHops takes a slice of hop pointers, the annotation data
 // mapping ip addresses to metadata and a timestamp. It will then use
 // these to attach the appropriate metadata to the PT hops.
-func AnnotatePTHops(hops []*schema.ParisTracerouteHop, annotationData map[string]schema.MetaData, timestamp time.Time) {
+func AnnotatePTHops(hops []*schema.ParisTracerouteHop, annotationData map[string]MetaData, timestamp time.Time) {
 	if annotationData == nil {
 		return
 	}
@@ -156,8 +152,8 @@ func AnnotatePTHops(hops []*schema.ParisTracerouteHop, annotationData map[string
 // and the associate timestamp. From those, it will create a slice of
 // requests to send to the annotation service, removing duplicates
 // along the way.
-func CreateRequestDataFromPTHops(hops []*schema.ParisTracerouteHop, timestamp time.Time) []schema.RequestData {
-	hopMap := map[string]schema.RequestData{}
+func CreateRequestDataFromPTHops(hops []*schema.ParisTracerouteHop, timestamp time.Time) []RequestData {
+	hopMap := map[string]RequestData{}
 	for _, hop := range hops {
 		if hop == nil {
 			metrics.AnnotationErrorCount.With(prometheus.
@@ -165,21 +161,21 @@ func CreateRequestDataFromPTHops(hops []*schema.ParisTracerouteHop, timestamp ti
 			continue
 		}
 		if hop.Src_ip != "" {
-			hopMap[hop.Src_ip] = schema.RequestData{hop.Src_ip, 0, timestamp}
+			hopMap[hop.Src_ip] = RequestData{hop.Src_ip, 0, timestamp}
 		} else {
 			metrics.AnnotationErrorCount.With(prometheus.
 				Labels{"source": "PT Hop was missing an IP!!!"}).Inc()
 		}
 
 		if hop.Dest_ip != "" {
-			hopMap[hop.Dest_ip] = schema.RequestData{hop.Dest_ip, 0, timestamp}
+			hopMap[hop.Dest_ip] = RequestData{hop.Dest_ip, 0, timestamp}
 		} else {
 			metrics.AnnotationErrorCount.With(prometheus.
 				Labels{"source": "PT Hop was missing an IP!!!"}).Inc()
 		}
 	}
 
-	requestSlice := make([]schema.RequestData, 0, len(hopMap))
+	requestSlice := make([]RequestData, 0, len(hopMap))
 	for _, req := range hopMap {
 		requestSlice = append(requestSlice, req)
 	}
@@ -215,13 +211,14 @@ func AddMetaDataPTHop(hop *schema.ParisTracerouteHop, timestamp time.Time) {
 			Labels{"source": "PT Hop had no dest_ip!"}).Inc()
 	}
 }
+*/
 
 // GetAndInsertGeolocationIPStruct takes a NON-NIL pointer to a
-// pre-allocated schema.GeolocationIP struct, an IP address, and a
+// pre-allocated GeolocationIP struct, an IP address, and a
 // timestamp. It will connect to the annotation service, get the
 // metadata, and insert the metadata into the reigion pointed to by
-// the schema.GeolocationIP pointer.
-func GetAndInsertGeolocationIPStruct(geo *schema.GeolocationIP, ip string, timestamp time.Time) {
+// the GeolocationIP pointer.
+func GetAndInsertGeolocationIPStruct(geo *GeolocationIP, ip string, timestamp time.Time) {
 	url := BaseURL + "ip_addr=" + url.QueryEscape(ip) +
 		"&since_epoch=" + strconv.FormatInt(timestamp.Unix(), 10)
 	annotationData := GetMetaData(url)
@@ -230,72 +227,11 @@ func GetAndInsertGeolocationIPStruct(geo *schema.GeolocationIP, ip string, times
 	}
 }
 
-// AddMetaDataNDTConnSpec takes a connection spec and a timestamp and
-// annotates the connection spec with metadata associated with each IP
-// Address. It will either sucessfully add the metadata or fail
-// silently and make no changes.
-func AddMetaDataNDTConnSpec(spec schema.Web100ValueMap, timestamp time.Time) {
-	// Time the response
-	timerStart := time.Now()
-	defer func(tStart time.Time) {
-		metrics.AnnotationTimeSummary.
-			With(prometheus.Labels{"test_type": "NDT"}).
-			Observe(float64(time.Since(tStart).Nanoseconds()))
-	}(timerStart)
-
-	GetAndInsertTwoSidedMetaIntoNDTConnSpec(spec, timestamp)
-}
-
-// GetAndInsertNDT takes a timestamp, an NDT connection spec, and a
-// string indicating whether it should get the metadata for the client
-// end or the server end of the connection. It will either insert the
-// data into the connection spec or silently fail.
-func GetAndInsertMetaIntoNDTConnSpec(side string, spec schema.Web100ValueMap, timestamp time.Time) {
-	ip, ok := spec.GetString([]string{side + "_ip"})
-	if ok {
-		url := BaseURL + "ip_addr=" + url.QueryEscape(ip) +
-			"&since_epoch=" + strconv.FormatInt(timestamp.Unix(), 10)
-		annotationData := GetMetaData(url)
-		if annotationData != nil && annotationData.Geo != nil {
-			CopyStructToMap(annotationData.Geo, spec.Get(side+"_geolocation"))
-		} else {
-			metrics.AnnotationErrorCount.With(prometheus.
-				Labels{"source": "Couldn't get metadata for the " + side + " side."}).Inc()
-		}
-	}
-}
-
-// CopyStructToMap takes a POINTER to an arbitrary struct and copies
-// it's fields into a value map. It will also make fields entirely
-// lower case, for convienece when working with exported structs. Also,
-// NEVER pass in something that is not a pointer to a struct, as this
-// will cause a panic.
-func CopyStructToMap(sourceStruct interface{}, destinationMap map[string]bigquery.Value) {
-	structToCopy := reflect.ValueOf(sourceStruct).Elem()
-	typeOfStruct := structToCopy.Type()
-	for i := 0; i < typeOfStruct.NumField(); i++ {
-		v := structToCopy.Field(i).Interface()
-		switch t := v.(type) {
-		case string:
-			if t == "" {
-				continue
-			}
-		case int64:
-			if t == 0 {
-				continue
-			}
-		}
-		destinationMap[strings.ToLower(typeOfStruct.Field(i).Name)] = v
-
-	}
-
-}
-
 // GetMetaData combines the functionality of QueryAnnotationService and
 // ParseJSONMetaDataResponse to query the annotator service and return
 // the corresponding MetaData if it can, or a nil pointer if it
 // encounters any error and cannot get the data for any reason
-func GetMetaData(url string) *schema.MetaData {
+func GetMetaData(url string) *MetaData {
 	// Query the service and grab the response safely
 	annotatorResponse, err := QueryAnnotationService(url)
 	if err != nil {
@@ -346,8 +282,8 @@ func QueryAnnotationService(url string) ([]byte, error) {
 // the JSON from the annotator service and parses it into a MetaData
 // struct, for easy manipulation. It returns a pointer to the struct on
 // success and an error if an error occurs.
-func ParseJSONMetaDataResponse(jsonBuffer []byte) (*schema.MetaData, error) {
-	parsedJSON := &schema.MetaData{}
+func ParseJSONMetaDataResponse(jsonBuffer []byte) (*MetaData, error) {
+	parsedJSON := &MetaData{}
 	err := json.Unmarshal(jsonBuffer, parsedJSON)
 	if err != nil {
 		return nil, err
@@ -355,59 +291,12 @@ func ParseJSONMetaDataResponse(jsonBuffer []byte) (*schema.MetaData, error) {
 	return parsedJSON, nil
 }
 
-// GetAndInsertTwoSidedMetaIntoNDTConnSpec takes a timestamp and an
-// NDT connection spec. It will either insert the data into the
-// connection spec or silently fail.
-func GetAndInsertTwoSidedMetaIntoNDTConnSpec(spec schema.Web100ValueMap, timestamp time.Time) {
-	// TODO(JM): Make metrics for sok and cok failures. And double check metrics for cleanliness.
-	cip, cok := spec.GetString([]string{"client_ip"})
-	sip, sok := spec.GetString([]string{"server_ip"})
-	reqData := []schema.RequestData{}
-	if cok {
-		reqData = append(reqData, schema.RequestData{IP: cip, Timestamp: timestamp})
-	} else {
-		metrics.AnnotationErrorCount.With(prometheus.
-			Labels{"source": "Missing client side IP."}).Inc()
-	}
-	if sok {
-		reqData = append(reqData, schema.RequestData{IP: sip, Timestamp: timestamp})
-	} else {
-		metrics.AnnotationErrorCount.With(prometheus.
-			Labels{"source": "Missing server side IP."}).Inc()
-	}
-	if cok || sok {
-		annotationDataMap := GetBatchMetaData(BatchURL, reqData)
-		// TODO(JM): Revisit decision to use base36 for
-		// encoding, rather than base64. (It had to do with
-		// library support.)
-		timeString := strconv.FormatInt(timestamp.Unix(), 36)
-		if cok {
-			if data, ok := annotationDataMap[cip+timeString]; ok && data.Geo != nil {
-				CopyStructToMap(data.Geo, spec.Get("client_geolocation"))
-			} else {
-				metrics.AnnotationErrorCount.With(prometheus.
-					Labels{"source": "Couldn't get metadata for the client side."}).Inc()
-			}
-		}
-		if sok {
-			if data, ok := annotationDataMap[sip+timeString]; ok && data.Geo != nil {
-				CopyStructToMap(data.Geo, spec.Get("server_geolocation"))
-			} else {
-				metrics.AnnotationErrorCount.With(prometheus.
-					Labels{"source": "Couldn't get metadata for the server side."}).Inc()
-			}
-
-		}
-	}
-
-}
-
 // GetBatchMetaData combines the functionality of
 // BatchQueryAnnotationService and BatchParseJSONMetaDataResponse to
 // query the annotator service and return the corresponding map of
-// ip-timestamp strings to schema.MetaData structs, or a nil map if it
+// ip-timestamp strings to MetaData structs, or a nil map if it
 // encounters any error and cannot get the data for any reason
-func GetBatchMetaData(url string, data []schema.RequestData) map[string]schema.MetaData {
+func GetBatchMetaData(url string, data []RequestData) map[string]MetaData {
 	// Query the service and grab the response safely
 	annotatorResponse, err := BatchQueryAnnotationService(url, data)
 	if err != nil {
@@ -429,30 +318,35 @@ func GetBatchMetaData(url string, data []schema.RequestData) map[string]schema.M
 }
 
 // BatchQueryAnnotationService takes a url to POST the request to and
-// a slice of schema.RequestDatas to be sent in the body in a JSON
+// a slice of RequestDatas to be sent in the body in a JSON
 // format. It will copy the response into a []byte and return it to
 // the user, returning an error if any occurs
-func BatchQueryAnnotationService(url string, data []schema.RequestData) ([]byte, error) {
+// TODO(gfr) Should pass the annotator's request context through and use it here.
+func BatchQueryAnnotationService(url string, data []RequestData) ([]byte, error) {
 	encodedData, err := json.Marshal(data)
 	if err != nil {
 		metrics.AnnotationErrorCount.
 			With(prometheus.Labels{"source": "Couldn't Marshal Data"}).Inc()
 		return nil, err
 	}
-	// Make the actual request
-	resp, err := http.Post(url, "raw", bytes.NewReader(encodedData))
 
+	var netClient = &http.Client{
+		Timeout: time.Second,
+	}
+
+	// Make the actual request
+	resp, err := netClient.Post(url, "raw", bytes.NewReader(encodedData))
 	// Catch http errors
 	if err != nil {
 		metrics.AnnotationErrorCount.
-			With(prometheus.Labels{"source": "Request to Annotator failed"}).Inc()
+			With(prometheus.Labels{"source": err.Error()}).Inc()
 		return nil, err
 	}
 
 	// Catch errors reported by the service
 	if resp.StatusCode != http.StatusOK {
 		metrics.AnnotationErrorCount.
-			With(prometheus.Labels{"source": "Webserver gave non-ok response"}).Inc()
+			With(prometheus.Labels{"source": http.StatusText(resp.StatusCode)}).Inc()
 		return nil, errors.New("URL:" + url + " gave response code " + resp.Status)
 	}
 	defer resp.Body.Close()
@@ -463,11 +357,11 @@ func BatchQueryAnnotationService(url string, data []schema.RequestData) ([]byte,
 
 // BatchParseJSONMetaDataResponse takes a byte slice containing the
 // text of the JSON from the annoator service's batch request endpoint
-// and parses it into a map of strings to schema.MetaData structs, for
+// and parses it into a map of strings to MetaData structs, for
 // easy manipulation. It returns a pointer to the struct on success
 // and an error if one occurs.
-func BatchParseJSONMetaDataResponse(jsonBuffer []byte) (map[string]schema.MetaData, error) {
-	parsedJSON := make(map[string]schema.MetaData)
+func BatchParseJSONMetaDataResponse(jsonBuffer []byte) (map[string]MetaData, error) {
+	parsedJSON := make(map[string]MetaData)
 	err := json.Unmarshal(jsonBuffer, &parsedJSON)
 	if err != nil {
 		return nil, err
